feat(intex): widen the series axis to the packed composite id
Series identity stops being the worldwide day and becomes SeriesId end to end:
the registry key and its enumeration, IssuanceParams, the settle counter, the
bin-index values, the authorized-settler and mine-seq key preimages, the settled
token id, the proof-of-work preimage and the read-only precompile. The money
axis keeps its u32 day, so distribute, the proceeds fan-in and the creator
payout are untouched.

Two things the type change forced out into the open:

The proceeds fan-in was armed with the series id although every credit arrives
per day, so the two keys would have diverged the moment a day produced more than
one series, stranding the pot. It is armed with the day now.

The certified-contributor path decided "has this day been issued" by probing the
series ledger at the day's own id. That answer disappears when the id stops
being the day, and it feeds a fatal-on-read version check. The registry now
counts series per day and answers from that, which stays correct once a day
derives several series.

The off-chain miner moves in the same commit: the hash preimage is shared with
the precompile, so a split would silently reject every mined nonce. The gem
position id follows too, since merchant gems burn Intex by series id.

+++ contracts/precompiles/src/IGemFactory.sol
@@ -5,7 +5,7 @@ interface IGemFactory {
     /// @notice Park the caller's Intex series `sourceIntexId` (burning `amount`
     ///         units via IntexNFT1155) and mint a GemPosition NFT to the caller.
     ///         Returns the new `positionId`.
-    function mintGemPosition(uint32 sourceIntexId, uint256 amount) external returns (uint256 positionId);
+    function mintGemPosition(uint64 sourceIntexId, uint256 amount) external returns (uint256 positionId);
     /// @notice Issue one Merchant gem to `owner`, draining the position's
     ///         capacity. Only the position's merchant (the caller) may call.
     function mintMerchantGem(uint256 positionId, address owner, uint256 gemLoad) external returns (uint256 gemId);
