+++ crates/core/intexfactory/src/sol_ext.rs
@@ -9,7 +9,7 @@ sol! {
     interface IOriginRouter {
         struct IssuanceInstructionsParams {
             uint32 dstChainId;
-            uint32 seriesId;
+            uint64 seriesId;
             uint32 worldwideDay;
             uint32 issuedIntexCount;
             uint128 promisLoadMinor;
@@ -28,9 +28,9 @@ sol! {
         function sendIssuanceInstructions(IssuanceInstructionsParams calldata params)
             external payable returns (bytes32 sendId);
 
-        function sendMarkQualified(uint32 seriesId) external payable returns (bytes32 sendId);
+        function sendMarkQualified(uint64 seriesId) external payable returns (bytes32 sendId);
 
-        function sendMarkCalled(uint32 seriesId) external payable returns (bytes32 sendId);
+        function sendMarkCalled(uint64 seriesId) external payable returns (bytes32 sendId);
     }
 
     #[sol(alloy_sol_types = alloy_sol_types)]
@@ -55,7 +55,7 @@ sol! {
         }
 
         struct CreateSeriesParams {
-            uint32 seriesId;
+            uint64 seriesId;
             uint32 worldwideDay;
             uint16 issuanceCurrency;
             uint16 referenceCurrency;
@@ -69,9 +69,9 @@ sol! {
 
         function createSeries(CreateSeriesParams params) external;
         function balanceOf(address account, uint256 id) external view returns (uint256);
-        function settle(uint32 seriesId, address from, address to, uint256 amount) external;
-        function burnSettled(address holder, uint32 seriesId, uint256 amount) external;
-        function markQualified(uint32 seriesId) external;
-        function markCalled(uint32 seriesId) external;
+        function settle(uint64 seriesId, address from, address to, uint256 amount) external;
+        function burnSettled(address holder, uint64 seriesId, uint256 amount) external;
+        function markQualified(uint64 seriesId) external;
+        function markCalled(uint64 seriesId) external;
     }
 }
